Fix: reviewController 이름 -> sidebarController로 변경 및 적용(Page - Controller 명 일치화)
